+++ app/src/main/java/cl/inacap/tdis08/sapo/captivemonitor/model/TankState.java
@@ -2,17 +2,17 @@ package cl.inacap.tdis08.sapo.captivemonitor.model;
 
 public class TankState {
 
-    private int soilHumidity;
-    private int soilTemperature;
-    private int roomHumidity;
-    private int roomTemperature;
-    private int waterLevel;
-    private int waterTemperature;
+    private double soilHumidity;
+    private double soilTemperature;
+    private double roomHumidity;
+    private double roomTemperature;
+    private double waterLevel;
+    private double waterTemperature;
 
     public TankState() {
     }
 
-    public TankState(int soilHumidity, int soilTemperature, int roomHumidity, int roomTemperature, int waterLevel, int waterTemperature) {
+    public TankState(double soilHumidity, double soilTemperature, double roomHumidity, double roomTemperature, double waterLevel, double waterTemperature) {
         this.soilHumidity = soilHumidity;
         this.soilTemperature = soilTemperature;
         this.roomHumidity = roomHumidity;
@@ -21,51 +21,51 @@ public class TankState {
         this.waterTemperature = waterTemperature;
     }
 
-    public int getSoilHumidity() {
+    public double getSoilHumidity() {
         return soilHumidity;
     }
 
-    public void setSoilHumidity(int soilHumidity) {
+    public void setSoilHumidity(double soilHumidity) {
         this.soilHumidity = soilHumidity;
     }
 
-    public int getSoilTemperature() {
+    public double getSoilTemperature() {
         return soilTemperature;
     }
 
-    public void setSoilTemperature(int soilTemperature) {
+    public void setSoilTemperature(double soilTemperature) {
         this.soilTemperature = soilTemperature;
     }
 
-    public int getRoomHumidity() {
+    public double getRoomHumidity() {
         return roomHumidity;
     }
 
-    public void setRoomHumidity(int roomHumidity) {
+    public void setRoomHumidity(double roomHumidity) {
         this.roomHumidity = roomHumidity;
     }
 
-    public int getRoomTemperature() {
+    public double getRoomTemperature() {
         return roomTemperature;
     }
 
-    public void setRoomTemperature(int roomTemperature) {
+    public void setRoomTemperature(double roomTemperature) {
         this.roomTemperature = roomTemperature;
     }
 
-    public int getWaterLevel() {
+    public double getWaterLevel() {
         return waterLevel;
     }
 
-    public void setWaterLevel(int waterLevel) {
+    public void setWaterLevel(double waterLevel) {
         this.waterLevel = waterLevel;
     }
 
-    public int getWaterTemperature() {
+    public double getWaterTemperature() {
         return waterTemperature;
     }
 
-    public void setWaterTemperature(int waterTemperature) {
+    public void setWaterTemperature(double waterTemperature) {
         this.waterTemperature = waterTemperature;
     }
