+++ app/src/main/java/cl/inacap/tdis08/sapo/captivemonitor/util/Presets.java
@@ -15,19 +15,19 @@ public class Presets {
     static {
         presets = new HashMap<>();
         presets.put(Species.FISH,
-                new TankPreset(new int[][]
+                new TankPreset(new double[][]
                         {{0,0},{0,0},{0,0},{0,0},{70,95},{20,23}}
                 ));
         presets.put(Species.FROG,
-                new TankPreset(new int[][]
+                new TankPreset(new double[][]
                         {{0,0},{0,0},{60,90},{24,30},{0,0},{0,0}}
                 ));
         presets.put(Species.TURTLE,
-                new TankPreset(new int[][]
+                new TankPreset(new double[][]
                         {{0,0},{29,32},{60,80},{24,29},{70,95},{22,25}}
                 ));
         presets.put(Species.LIZARD,
-                new TankPreset(new int[][]
+                new TankPreset(new double[][]
                         {{10,40},{28,32},{30,40},{22,26},{0,0},{0,0}}
                 ));
     }
